requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1
dj-database-url>=2.1.0
gunicorn>=21.2.0
whitenoise>=6.6.0
